Player Data Update
Added ability to create a user account and keep data.
Automatic saving every 5 minutes and on close
Created a nice home page to allow players to create accounts and sign
in.
Automatic reconnection in the background when the connection is lost.
Rearranged some files
Added envrionment variables (will need to update .gitignore)